Update file-uploads.blade.php
In order to make the test pass we need to specify a disk to fake and use in the test and in the component respectively. Otherwise the test doesn't work because Livewire creates and replaces the default disk for a new fake disk (temp-for-tests) under the hood.

The internal code of Livewire could be changed to fix this issue; but in the meantime this provides a more detailed  and working example (I took this example from Livewire tests).