Update actions.blade.php
I just corrected one example but it should be done for all the subsection. 

If you have a look to the Laravel document, each subsection has a # symbol and when you click on each subsection suddenly the url will be changed accordingly. This is usually needed when you are scrolling the document and want to reference to that somewhere in a blog.

I don't see such feature in this Livewire document and it is needed.