Add a note for keys in paginated results
The problem with using `$loop->index` as a key in a paginated result is the fact it might duplicate. 
For example: Lets say we have somewhere between 11 to 20 items, paginated with 10 items per page.
On page 2, the first  view-item component (for item 11)  will have `:key=0` which is the same as the first item from page 1. This can cause the wrong data to appear in the component.