Clarification on validateOnly when using arrays
I assumed that I could validate the items of the array in the same "validateOnly" call I was using to validate an array until I realized that I had to be more precise and indicate that in the property name parameter of the validateOnly method. I would propose adding this small clarification if you believe someone else could make the same mistake.