Add docs for max_upload_time
We are running an app with ~1900 users. Since we implemented an upload with Livewire we occasionally got complaints that the upload doesn't work. Files that are uploaded on our app usually have a size between 50 to 75Mb. I finally figured out that the failed upload is only occurring on slow internet connections. It took a while until I found that setting, therefore I made this PR so others may be able to find this easier :)